% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial utilities.R
\name{categorise_files}
\alias{categorise_files}
\title{Categorise files for processing with `NEMO_ERSEM`}
\usage{
categorise_files(dir, recursive = TRUE)
}
\arguments{
\item{dir}{A path string to the directory containing NEMO-ERSEM netcdf files.}

\item{recursive}{T/F whether to process files in all sub-directories.}
}
\value{
A dataframe that can be passed to `NEMO_ERSEM` using pmap.
}
\description{
This function takes a directory, recursively lists the file names, and extracts meaningful meta-data
}
\seealso{
Other NEMO-ERSEM spatial tools: 
\code{\link{calculate_depth_share}()},
\code{\link{calculate_proximity_weight}()},
\code{\link{netcdf_scheme_helpers}},
\code{\link{scheme_column}()},
\code{\link{scheme_interp_slice}()},
\code{\link{scheme_strathE2E}()},
\code{\link{voronoi_grid}()},
\code{\link{xyindex_to_nindex}()}
}
\concept{NEMO-ERSEM spatial tools}
