% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit conversion functions.R
\name{Convert-units}
\alias{Convert-units}
\alias{micro_to_full}
\alias{full_to_micro}
\alias{micro_to_milli}
\alias{milli_to_micro}
\alias{milli_to_full}
\alias{full_to_milli}
\alias{l_to_m3}
\alias{m3_to_l}
\alias{sec_to_day}
\alias{day_to_sec}
\alias{shortwave_to_einstein}
\title{Helper functions for common unit conversions}
\usage{
micro_to_full(data)

full_to_micro(data)

micro_to_milli(data)

milli_to_micro(data)

milli_to_full(data)

full_to_milli(data)

l_to_m3(data)

m3_to_l(data)

sec_to_day(data)

day_to_sec(data)

shortwave_to_einstein(data)
}
\arguments{
\item{data}{A numeric vector.}
}
\value{
A numeric vector following unit conversion.
}
\description{
These functions are named to make it intuitive to switch between scales, with pairs existing for
different directions of conversion.
}
\details{
Though some of these functions do the same thing numerically, having multiple names which match the desired
conversion should reduce friction when developing code. This should also help avoid typos when it comes to
0s, and also stop you needing to constantly look up conversion factors!

The `shortwave_to_einsteins` function is not exact and is based on a couple of assumptions to get from watts
to einsteins per m^2 per day. It expects an input of total Watts per day per m^2.
\itemize{
 \item{The proportion of PAR from shortwave downward radiation is 0.43.}
 \item{The energy content of a mole of photons is 217.503 kJ, taken from midband radiation (550nm).}}
}
