% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inspecting.R
\name{decadal}
\alias{decadal}
\title{Prepare for Averaging by Decade}
\usage{
decadal(saved)
}
\arguments{
\item{saved}{A dataframe containing a summarised month from NEMO-ERSEM model outputs.}
}
\value{
A dataframe containing a summarised month of NEMO-ERSEM output, gaining a decade column, and dropping columns
which aren't needed for spatial maps.
}
\description{
This function cleans the saved NEMO-ERSEM monthly summaries, for averaging into decades.
}
\seealso{
Other NEMO-ERSEM averages: 
\code{\link{NE_decadal_summary}()}
}
\concept{NEMO-ERSEM averages}
