% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extractor functions MA.R
\name{extractors_MA}
\alias{extractors_MA}
\alias{get_so_abs_slabR}
\alias{get_thetao_con_slabR}
\alias{get_R1_slabR}
\alias{get_O2_slabR}
\alias{get_N4_slabR}
\alias{get_N3_slabR}
\alias{get_RP_slabR}
\alias{get_B1_slabR}
\alias{get_P1_slabR}
\alias{get_P234_n_slabR}
\alias{get_difvho_slabR}
\alias{get_wo_slabR}
\alias{get_vo_slabR}
\alias{get_uo_slabR}
\title{Extract summaries from Mission Atlantic NEMO-ERSEM arrays using slabR}
\usage{
get_so_abs_slabR(
  path,
  file,
  scheme,
  start = c(1, 1, 1, 1),
  count = c(-1, -1, -1, -1),
  collapse_days = TRUE,
  ...
)

get_thetao_con_slabR(
  path,
  file,
  scheme,
  start = c(1, 1, 1, 1),
  count = c(-1, -1, -1, -1),
  collapse_days = TRUE,
  ...
)

get_R1_slabR(
  path,
  file,
  scheme,
  start = c(1, 1, 1, 1),
  count = c(-1, -1, -1, -1),
  collapse_days = TRUE,
  ...
)

get_O2_slabR(
  path,
  file,
  scheme,
  start = c(1, 1, 1, 1),
  count = c(-1, -1, -1, -1),
  collapse_days = TRUE,
  ...
)

get_N4_slabR(
  path,
  file,
  scheme,
  start = c(1, 1, 1, 1),
  count = c(-1, -1, -1, -1),
  collapse_days = TRUE,
  ...
)

get_N3_slabR(
  path,
  file,
  scheme,
  start = c(1, 1, 1, 1),
  count = c(-1, -1, -1, -1),
  collapse_days = TRUE,
  ...
)

get_RP_slabR(
  path,
  file,
  scheme,
  start = c(1, 1, 1, 1),
  count = c(-1, -1, -1, -1),
  collapse_days = TRUE,
  ...
)

get_B1_slabR(
  path,
  file,
  scheme,
  start = c(1, 1, 1, 1),
  count = c(-1, -1, -1, -1),
  collapse_days = TRUE,
  ...
)

get_P1_slabR(
  path,
  file,
  scheme,
  start = c(1, 1, 1, 1),
  count = c(-1, -1, -1, -1),
  collapse_days = TRUE,
  ...
)

get_P234_n_slabR(
  path,
  file,
  scheme,
  start = c(1, 1, 1, 1),
  count = c(-1, -1, -1, -1),
  collapse_days = TRUE,
  ...
)

get_difvho_slabR(
  path,
  file,
  scheme,
  start = c(1, 1, 1, 1),
  count = c(-1, -1, -1, -1),
  collapse_days = TRUE,
  ...
)

get_wo_slabR(
  path,
  file,
  scheme,
  start = c(1, 1, 1, 1),
  count = c(-1, -1, -1, -1),
  collapse_days = TRUE,
  ...
)

get_vo_slabR(
  path,
  file,
  scheme,
  start = c(1, 1, 1, 1),
  count = c(-1, -1, -1, -1),
  collapse_days = TRUE,
  ...
)

get_uo_slabR(
  path,
  file,
  scheme,
  start = c(1, 1, 1, 1),
  count = c(-1, -1, -1, -1),
  collapse_days = TRUE,
  ...
)
}
\arguments{
\item{path}{the path to the NEMO-ERSEM model outputs.}

\item{file}{the name of a netcdf file containing the title variables.}

\item{scheme}{a summary scheme as expected by \code{array_w_mean()}.}

\item{start}{an optional vector of indices to start subsetting at. See ncdf4 documentation.}

\item{count}{an optional vector of steps to subset along. See ncdf4 documentation.}

\item{collapse_days}{a logical switch setting whether to extract all days within a single netcdf file, or to take the mean across days.}

\item{...}{soaks up unused function arguments passed by the wrapper functions handling file architecture.}
}
\value{
A matrix with a column of group averages per title variable for a single day.
}
\description{
These functions read in target variables from NEMO-ERSEM model outputs and return weighted averages
according to a summary scheme.
}
\details{
Each variable of interest in a netcdf file is imported, only reading within an x/y window specified
with \code{start} and \code{count}. The values are then passed to \code{array_w_mean()} to summarise according to a scheme.

As the Mission Atlantic data files contain multiple time steps, these functions calculate the mean monthly
array BEFORE using \code{array_w_mean()} for the spatial operation.

Each model output file type contains a different variable, extracted by the relevant function variant:\tabular{ll}{
   File type         \tab Variables \cr
   thetao_con \tab Temperature. Celsius \cr
   so_abs \tab Salinity. \cr
   uo \tab Zonal currents. m/s \cr
   vo \tab Meridional currents. m/s \cr
   wo \tab Vertical velocity. \cr
   difvho \tab Vertical diffusivity. m2/s \cr
   R1_n \tab Dissolved organic nitrogen. mmolN.m-3 \cr
   O2_o \tab Oxygen. mmol.m-3 \cr
   N4_n \tab Ammonium. mmolN.m-3 \cr
   N3_n \tab Nitrate. mmolN.m-3 \cr
   RP_n_result \tab Detritus (Particulate organic nitrogen). mmolN.m-3 \cr
   B1_n \tab Bacterial nitrogen. mmolN.m-3 \cr
   P1_n \tab Diatom nitrogen. mmolN.m-3 \cr
   P234_n_result \tab Other phytoplankton nitrogen. mmolN.m-3 \cr
}


Some function variants have different arguments:

The Mission Atlantic dataset was interpolated onto a regular lat-lon and depth grid. There is no need for any grid_w arguments in contrast to the nemomedusR package.
}
\seealso{
Other NEMO-ERSEM variable extractors: 
\code{\link{NEMO_ERSEM}()},
\code{\link{extractors_1D}},
\code{\link{extractors_slabR}},
\code{\link{get_spatial}()},
\code{\link{process_array}()}
}
\concept{NEMO-ERSEM variable extractors}
