% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial utilities.R
\name{get_spatial}
\alias{get_spatial}
\title{Get Latitudes, Longitudes, & Depths From NEMO-ERSEM Model Outputs}
\usage{
get_spatial(file, depthvar = NULL, depthdim = NULL)
}
\arguments{
\item{file}{The full name of a netcdf file.}

\item{depthvar}{The name of the depth dimension in a netcdf file (when extracting values from dimension metadata).}
}
\value{
A list of three elements:
\itemize{
 \item{\emph{nc_lat -}}{ A matrix of latitudes which maps onto the first and second dimension of a NEMO-ERSEM array.}
 \item{\emph{nc_lon -}}{ A matrix of longitudes which maps onto the first and second dimension of a NEMO-ERSEM array.}
 \item{\emph{nc_depth -}}{ A vector of depths which match the third dimension of a NEMO-ERSEM array.}
 }
}
\description{
This function gets the latitudes, longitudes, and depths which define the spatial location of points in an array of NEMO-ERSEM outputs.
}
\details{
Each variable of interest in the netcdf file is imported, and then collected into a list.
}
\seealso{
Other NEMO-ERSEM variable extractors: 
\code{\link{NEMO_ERSEM}()},
\code{\link{extractors_1D}},
\code{\link{extractors_MA}},
\code{\link{extractors_slabR}},
\code{\link{process_array}()}
}
\concept{NEMO-ERSEM variable extractors}
