% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial utilities.R
\name{NE_boundary_summary}
\alias{NE_boundary_summary}
\title{Summarise NEMO-ERSEM Output into Time Series Along Transects}
\usage{
NE_boundary_summary(saved, transects, vars = c("NO3", "NH4", "Detritus"))
}
\arguments{
\item{saved}{A dataframe containing a summarised month from NEMO-ERSEM model outputs.}

\item{transects}{A dataframe containing the labelled transects along the model domain boundaries.}

\item{vars}{A character vector containing the column names to be summarised for boundary conditions. Defaults to the targets for StrathE2E}
}
\value{
A list containing two summaries.
\itemize{
 \item{Element 1}{A dataframe containing the total water exchanged between model compartments.}
 \item{Element 2}{A dataframe containing the flow-weighted boundary conditions around the model domain.}}
}
\description{
This function averages NEMO-ERSEM summaries into time series for the target boundaries of StrathE2E.
}
\details{
The function subsets the NEMO-ERSEM grid according to the transects object provided. Water exchanges between
model compartments are totaled. The boundary conditions of the model domain for variables needed by
StrathE2E are summarised as a flow-weighted mean, applying the flow rate at each transect. As transects metadata is bound
to NEMO-ERSEM output, the function can handle daily or monthly summary objects, but the output will still be a monthly average
as needed by StrathE2E.
}
\seealso{
Other NEMO-MEDUSA averages: 
\code{\link{NE_volume_summary}()}
}
\concept{NEMO-MEDUSA averages}
