% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial utilities.R
\name{NE_decadal_summary}
\alias{NE_decadal_summary}
\title{Summarise NEMO-ERSEM Output into Decadal Grids
This function averages cleaned NEMO-ERSEM monthly summaries into decadal grids.}
\usage{
NE_decadal_summary(decade, dt)
}
\arguments{
\item{dt}{Switch for using either data.table or dplyr methods (TRUE/FALSE respectively)}

\item{saved}{A dataframe containing a summarised month from NEMO-ERSEM model outputs. It must contain the columns:
Longitude, Latitude, Decade, Month, Shore, and Depth.}
}
\value{
A dataframe containing a summarised decade of spatialy resolved NEMO-ERSEM outputs.
}
\description{
The function groups by all spatial variables (Longitude, Latitude, Depth, and Shore zone), and by decade and month.
The mean for every other variable is calculated within these groups.
}
\seealso{
Other NEMO-ERSEM averages: 
\code{\link{decadal}()}
}
\concept{NEMO-ERSEM averages}
