% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial utilities.R
\name{NE_volume_summary}
\alias{NE_volume_summary}
\title{Summarise NEMO-ERSEM Output into Time Series Within Model Compartments
This function averages NEMO-ERSEM monthly summaries into time series for each model compartment.}
\usage{
NE_volume_summary(saved, ice_threshold = 0, ice = FALSE)
}
\arguments{
\item{saved}{A dataframe containing a summarised month from NEMO-ERSEM model outputs.}

\item{ice_threshold}{A value between 0 and 1 defining ice free pixels.}

\item{ice}{A TRUE FALSE switch for whether the ice_mod files were included when extracting from NEMO-ERSEM.}
}
\value{
A dataframe containing a mean monthly time series of all target variables in NEMO-ERSEM outputs.
}
\description{
The function groups by model compartment (Depth and Shore zone) and time step (Month and Year).
The mean for every target variable is calculated within these groups.
}
\details{
The ice-threshold parameter controls how high the concentration of ice must be at a pixel to count as ice-affected.
The default is set to 0, so any pixels with ice are counted, but if you decide that ice-concentrations below 0.05
are meaningless for your purposes, you can set the threshold to have these pixels classed as ice free instead. This will
also drop the pixels from calculations of Ice_concentration, Ice_Thickness, and Snow_Thickness.
}
\seealso{
Other NEMO-MEDUSA averages: 
\code{\link{NE_boundary_summary}()}
}
\concept{NEMO-MEDUSA averages}
