% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nemoRsem-package.R
\docType{package}
\name{nemoRsem-package}
\alias{nemoRsem}
\alias{nemoRsem-package}
\title{nemoRsem: Tools for summarising NEMO-ERSEM model output}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package helps produce common spatial and temporal summaries from NEMO-ERSEM model output.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jack-h-laverick.github.io/nemoRsem/}
  \item \url{https:github.com/Jack-H-Laverick/nemoRsem}
  \item Report bugs at \url{https:github.com/Jack-H-Laverick/nemoRsem/issues}
}

}
\author{
\strong{Maintainer}: Jack H Laverick \email{jacklaverick@ymail.com} (\href{https://orcid.org/0000-0001-8829-2084}{ORCID})

}
\keyword{internal}
